show outputs in ui
