+++ client/src/models/testSuiteModels.ts
@@ -35,6 +35,7 @@ export interface Result {
   result_message?: string;
   created_at?: string;
   updated_at: string;
+  output_json?: string;
 }
 
 export interface TestInput {
